+++ CHANGELOG.md
@@ -11,3 +11,12 @@
 
 ## 0.0.4
   * Fix syntax.
+
+## 0.0.5
+  * Fix getting BrokenPipeError when calling singer.write_state.
+
+## 0.0.6
+  * Remove handling BrokenPipeError, because this is the expected behaviour in some situations.
+
+## 1.0.0
+  * Inhacement: Add missing keys to rows and assing a default value ('').